Exception handling

+++ src/main/java/com/sebastiao/dscommerce/services/ProductService.java
@@ -1,12 +1,18 @@
 package com.sebastiao.dscommerce.services;
 
+import com.fasterxml.jackson.databind.exc.InvalidDefinitionException;
 import com.sebastiao.dscommerce.dto.ProductDTO;
 import com.sebastiao.dscommerce.entities.Product;
 import com.sebastiao.dscommerce.repositories.ProductRepository;
+import com.sebastiao.dscommerce.services.exceptions.DatabaseException;
+import com.sebastiao.dscommerce.services.exceptions.ResourceNotFoundException;
+import jakarta.persistence.EntityNotFoundException;
 import org.springframework.beans.factory.annotation.Autowired;
+import org.springframework.dao.DataIntegrityViolationException;
 import org.springframework.data.domain.Page;
 import org.springframework.data.domain.Pageable;
 import org.springframework.stereotype.Service;
+import org.springframework.transaction.annotation.Propagation;
 import org.springframework.transaction.annotation.Transactional;
 
 import java.util.List;
@@ -26,7 +32,8 @@ public class ProductService {
 
     @Transactional(readOnly = true)
     public ProductDTO findById(Long id) {
-        Product product = repository.findById(id).get();
+        Product product = repository.findById(id).orElseThrow(
+                () -> new ResourceNotFoundException("Recurso não encontrado"));
         return new ProductDTO(product);
     }
 
@@ -40,15 +47,26 @@ public class ProductService {
 
     @Transactional
     public ProductDTO update(Long id, ProductDTO dto) {
-        Product obj = repository.getReferenceById(id);
-        copyDtoToEntity(dto, obj);
-        obj = repository.save(obj);
-        return new ProductDTO(obj);
+        try {
+            Product obj = repository.getReferenceById(id);
+            copyDtoToEntity(dto, obj);
+            obj = repository.save(obj);
+            return new ProductDTO(obj);
+        }catch (EntityNotFoundException e) {
+            throw new ResourceNotFoundException("Recurso não encontrado");
+        }
     }
 
-    @Transactional
+    @Transactional(propagation = Propagation.SUPPORTS)
     public void delete(Long id) {
-        repository.deleteById(id);
+        if(!repository.existsById(id)) {
+            throw new ResourceNotFoundException("Recurso não encontrado");
+        }
+        try {
+            repository.deleteById(id);
+        }catch (DataIntegrityViolationException e){
+            throw new DatabaseException("Falha de integridade referencial");
+        }
     }
 
     private void copyDtoToEntity(ProductDTO dto, Product obj) {
